Revert "Refactor (#6)"
This reverts commit 7915556a63ae4aee5a29cfdff41dbc43955f1e51.
