+++ Dockerfile
@@ -1,8 +1,11 @@
-FROM python:3.8-alpine3.12
+FROM alpine:3.12
+
+RUN apk add python3
 
 COPY ./start.sh /start.sh
-COPY ./martian_packets /martian_packets
+COPY ./main.py /main.py
 
 RUN chmod 0744 /start.sh
+RUN chmod 0744 /main.py
 
 CMD ./start.sh
